Add some comments, move AWDL functions to iOS-specific build tags
